+++ sentry/src/main/java/io/sentry/IFeedbackApi.java
@@ -2,7 +2,6 @@ package io.sentry;
 
 import io.sentry.protocol.Feedback;
 import io.sentry.protocol.SentryId;
-import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -21,14 +20,14 @@ public interface IFeedbackApi {
    * SentryFeedbackOptions#isUseShakeGesture()}. Only supported on Android; no-op on other
    * platforms.
    */
-  void enableFeedbackOnShake();
+  void enableOnShake();
 
   /**
    * Disables showing the feedback form when a shake gesture is detected, overriding {@link
    * SentryFeedbackOptions#isUseShakeGesture()}. Only supported on Android; no-op on other
    * platforms.
    */
-  void disableFeedbackOnShake();
+  void disableOnShake();
 
   /**
    * Whether showing the feedback form on a shake gesture is currently enabled. Always {@code false}
@@ -36,8 +35,7 @@ public interface IFeedbackApi {
    *
    * @return true if the feedback form is shown when a shake gesture is detected
    */
-  @ApiStatus.Internal
-  boolean isFeedbackOnShakeEnabled();
+  boolean isOnShakeEnabled();
 
   @NotNull
   SentryId capture(final @NotNull Feedback feedback);